// scolaire-front/src/app/services/models/course-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CourseResponse {
  academieYearId?: number;
  archive?: boolean;
  classeId?: number;
  description?: string;
  halfYearlyId?: number;
  id?: number;
  name?: string;
  subjectId?: number
  classeName?: string;
  subjectName?: string;
  halfYearlyName?: string;
  academieName: string
}